requirements.txt
google-auth-oauthlib==1.0.0
google-api-python-client==2.108.0
openai==1.63.0
python-dotenv==1.0.0
pandas==2.1.3
